+++ control/src/cache/reconcile/tone.rs
@@ -1,9 +1,26 @@
 use super::calculate;
+use crate::cache::display::{AltMenuScreen, Screen, TonePosition};
 use crate::cache::mapping::AttributeIdentifier;
 use crate::Store;
 
 impl Store {
     pub fn reconcile_tone(&mut self) {
+        if self.input.button.pressed && self.input.tone.active() {
+            if self.input.tone.value() > 0.5 {
+                self.cache.options.filter_feedback = false;
+                self.cache.display.set_alt_menu(Screen::AltMenu(
+                    0,
+                    AltMenuScreen::TonePosition(TonePosition::Volume),
+                ));
+            } else {
+                self.cache.options.filter_feedback = true;
+                self.cache.display.set_alt_menu(Screen::AltMenu(
+                    0,
+                    AltMenuScreen::TonePosition(TonePosition::Feedback),
+                ));
+            }
+        }
+
         self.cache.attributes.tone = calculate(
             self.input.tone.value(),
             self.control_value_for_attribute(AttributeIdentifier::Tone),