+++ CHANGELOG.md
@@ -11,8 +11,8 @@ backwards compatibility.
 * Finalize user manual's front page.
 * Trigger flutter in sporadic pulses.
 * Optimize wow and flutter algorithms.
-* Control pre-amp mode using alt button and potentiometer.
-* Control speed range using alt button and potentiometer.
+* Replace DIP switch with alt button and potentiomenter for:
+  pre-amp mode, speed range, tone filter position.
 
 ## 0.5.0